pyproject.toml
[project]
name = "excel-rag-system"
version = "0.1.0"
description = "Excel RAG System with LangGraph and Gemini"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "tantivy>=0.22.0",
    "openpyxl>=3.1.0",
    "numpy>=1.24.0",
    "langchain-google-genai>=1.0.0",
    "langgraph>=0.0.20",
    "langchain-core>=0.1.0",
    "python-dotenv>=1.0.0",
    "langchain-chroma>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]
ignore = ["E501"]
